src/test/java/com/safetynet/alerts/web/serialization/dao/OtherResidentsDaoTest.java
package com.safetynet.alerts.web.serialization.dao;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.io.IOException;
import java.io.StringWriter;
import java.util.ArrayList;
import java.util.List;

import org.junit.jupiter.api.Test;

import com.fasterxml.jackson.core.JsonFactory;
import com.fasterxml.jackson.core.JsonGenerator;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.safetynet.alerts.web.model.Person;

public class OtherResidentsDaoTest {

    /**
     * SomeJavadoc.
     * Test 'Serialize' method for 'OtherResidentsDao'.
     */
    @Test
    void testSerialize() {
        OtherResidentsDao otherResidentsDao = new OtherResidentsDao(Person.class);

        Person person = new Person();
        person.setFirstName("John");
        person.setLastName("Doe");

        try {
            ObjectMapper objectMapper = new ObjectMapper();
            StringWriter writer = new StringWriter();
            JsonGenerator jsonGenerator = new JsonFactory().createGenerator(writer);

            otherResidentsDao.serialize(person, jsonGenerator, objectMapper.getSerializerProvider());
            jsonGenerator.flush();

            String expectedJson = "{\"firstName\":\"John\",\"lastName\":\"Doe\"}";
            String actualJson = writer.toString();

            assertEquals(expectedJson, actualJson);
        } catch (IOException e) {
            e.printStackTrace();
        }
    }

    /**
     * SomeJavadoc.
     * Test 'SerializeList' method for 'OtherResidentsDao'.
     */
    @Test
    void testSerializeList() {
        OtherResidentsDao otherResidentsDao = new OtherResidentsDao(Person.class);

        List<Person> persons = new ArrayList<>();
        Person person1 = new Person();
        person1.setFirstName("John");
        person1.setLastName("Doe");

        Person person2 = new Person();
        person2.setFirstName("Jane");
        person2.setLastName("Smith");

        persons.add(person1);
        persons.add(person2);

        try {
            ObjectMapper objectMapper = new ObjectMapper();
            StringWriter writer = new StringWriter();
            JsonGenerator jsonGenerator = new JsonFactory().createGenerator(writer);

            otherResidentsDao.serializeList(persons, jsonGenerator, objectMapper.getSerializerProvider());
            jsonGenerator.flush();

            String expectedJson = "[{\"firstName\":\"John\",\"lastName\":\"Doe\"},{\"firstName\":\"Jane\",\"lastName\":\"Smith\"}]";
            String actualJson = writer.toString();

            assertEquals(expectedJson, actualJson);
        } catch (IOException e) {
            e.printStackTrace();
        }
    }
}
